perf(maintainability) :recycle: Maintainability improvments
modified index.js to use socketAndOther.js
